requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
geopandas==1.1.1
hdbscan==0.8.40
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.5.5.64
openslide-python==1.4.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pynndescent==0.5.13
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.19.3
scikit-learn==1.7.1
scipy==1.8.1
shapely==2.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
triton==2.3.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
xformers==0.0.27
